Move media picker files into ui package
