+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -80,8 +80,8 @@ import org.wordpress.android.ui.people.RoleChangeDialogFragment;
 import org.wordpress.android.ui.people.RoleSelectDialogFragment;
 import org.wordpress.android.ui.photopicker.PhotoPickerActivity;
 import org.wordpress.android.ui.photopicker.PhotoPickerFragment;
-import org.wordpress.android.ui.photopicker.mediapicker.MediaPickerActivity;
-import org.wordpress.android.ui.photopicker.mediapicker.MediaPickerFragment;
+import org.wordpress.android.ui.mediapicker.MediaPickerActivity;
+import org.wordpress.android.ui.mediapicker.MediaPickerFragment;
 import org.wordpress.android.ui.plans.PlanDetailsFragment;
 import org.wordpress.android.ui.plans.PlansActivity;
 import org.wordpress.android.ui.plans.PlansListAdapter;
