+++ WordPress/src/main/java/org/wordpress/android/modules/ViewModelModule.java
@@ -7,7 +7,7 @@ import org.wordpress.android.ui.JetpackRemoteInstallViewModel;
 import org.wordpress.android.ui.domains.DomainRegistrationMainViewModel;
 import org.wordpress.android.ui.main.MeViewModel;
 import org.wordpress.android.ui.photopicker.PhotoPickerViewModel;
-import org.wordpress.android.ui.photopicker.mediapicker.MediaPickerViewModel;
+import org.wordpress.android.ui.mediapicker.MediaPickerViewModel;
 import org.wordpress.android.ui.plans.PlansViewModel;
 import org.wordpress.android.ui.posts.EditPostPublishSettingsViewModel;
 import org.wordpress.android.ui.posts.PostListMainViewModel;
