+++ WordPress/src/main/java/org/wordpress/android/ui/mediapicker/MediaLoaderFactory.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.photopicker.mediapicker
+package org.wordpress.android.ui.mediapicker
 
 import javax.inject.Inject
