+++ WordPress/src/main/java/org/wordpress/android/ui/mediapicker/MediaType.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.photopicker.mediapicker
+package org.wordpress.android.ui.mediapicker
 
 enum class MediaType {
     IMAGE, VIDEO, AUDIO, DOCUMENT
