+++ WordPress/src/main/java/org/wordpress/android/ui/mediapicker/PhotoThumbnailViewHolder.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.photopicker.mediapicker
+package org.wordpress.android.ui.mediapicker
 
 import android.view.ViewGroup
 import android.widget.ImageView
