+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/PhotoPickerAdapter.kt
@@ -11,7 +11,7 @@ import org.wordpress.android.ui.photopicker.PhotoPickerUiItem.VideoItem
 import org.wordpress.android.util.image.ImageManager
 
 @Deprecated("This class is being refactored, if you implement any change, please also update " +
-        "{@link org.wordpress.android.ui.photopicker.mediapicker.MedaPickerAdapter}")
+        "{@link org.wordpress.android.ui.mediapicker.MedaPickerAdapter}")
 class PhotoPickerAdapter internal constructor(private val imageManager: ImageManager) : Adapter<ThumbnailViewHolder>() {
     private val thumbnailViewUtils = ThumbnailViewUtils(imageManager)
     private var mediaList = listOf<PhotoPickerUiItem>()
