+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/PhotoPickerAdapterDiffCallback.kt
@@ -5,7 +5,7 @@ import org.wordpress.android.ui.photopicker.PhotoPickerAdapterDiffCallback.Paylo
 import org.wordpress.android.ui.photopicker.PhotoPickerAdapterDiffCallback.Payload.SELECTION_CHANGE
 
 @Deprecated("This class is being refactored, if you implement any change, please also update " +
-        "{@link org.wordpress.android.ui.photopicker.mediapicker.MedaPickerAdapterDiffCallback}")
+        "{@link org.wordpress.android.ui.mediapicker.MedaPickerAdapterDiffCallback}")
 class PhotoPickerAdapterDiffCallback(
     private val oldItems: List<PhotoPickerUiItem>,
     private val updatedItems: List<PhotoPickerUiItem>
