+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/PhotoPickerItem.kt
@@ -3,7 +3,7 @@ package org.wordpress.android.ui.photopicker
 import org.wordpress.android.util.UriWrapper
 
 @Deprecated("This class is being refactored, if you implement any change, please also update " +
-        "{@link org.wordpress.android.ui.photopicker.mediapicker.MediaItem}")
+        "{@link org.wordpress.android.ui.mediapicker.MediaItem}")
 data class PhotoPickerItem(
     val id: Long = 0,
     val uri: UriWrapper? = null,
