+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/PhotoPickerUiItem.kt
@@ -5,7 +5,7 @@ import org.wordpress.android.ui.photopicker.PhotoPickerUiItem.Type.VIDEO
 import org.wordpress.android.util.UriWrapper
 
 @Deprecated("This class is being refactored, if you implement any change, please also update " +
-        "{@link org.wordpress.android.ui.photopicker.mediapicker.MedaPickerUiItem}")
+        "{@link org.wordpress.android.ui.mediapicker.MedaPickerUiItem}")
 sealed class PhotoPickerUiItem(
     val type: Type,
     open val id: Long,
