+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/PhotoPickerViewModel.kt
@@ -60,7 +60,7 @@ import javax.inject.Inject
 import javax.inject.Named
 
 @Deprecated("This class is being refactored, if you implement any change, please also update " +
-        "{@link org.wordpress.android.ui.photopicker.mediapicker.MediaPickerViewModel}")
+        "{@link org.wordpress.android.ui.mediapicker.MediaPickerViewModel}")
 class PhotoPickerViewModel @Inject constructor(
     @Named(UI_THREAD) private val mainDispatcher: CoroutineDispatcher,
     @Named(BG_THREAD) private val bgDispatcher: CoroutineDispatcher,
