+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/PhotoThumbnailViewHolder.kt
@@ -10,7 +10,7 @@ import org.wordpress.android.R
  */
 
 @Deprecated("This class is being refactored, if you implement any change, please also update " +
-        "{@link org.wordpress.android.ui.photopicker.mediapicker.PhotoThumbnailViewHolder}")
+        "{@link org.wordpress.android.ui.mediapicker.PhotoThumbnailViewHolder}")
 class PhotoThumbnailViewHolder(parent: ViewGroup, private val thumbnailViewUtils: ThumbnailViewUtils) :
         ThumbnailViewHolder(parent, R.layout.photo_picker_thumbnail) {
     private val imgThumbnail: ImageView = itemView.findViewById(R.id.image_thumbnail)
