+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/ThumbnailViewHolder.kt
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.RecyclerView.ViewHolder
  */
 
 @Deprecated("This class is being refactored, if you implement any change, please also update " +
-        "{@link org.wordpress.android.ui.photopicker.mediapicker.ThumbnailViewHolder}")
+        "{@link org.wordpress.android.ui.mediapicker.ThumbnailViewHolder}")
 open class ThumbnailViewHolder(parent: ViewGroup, layout: Int) : ViewHolder(
         LayoutInflater.from(parent.context)
                 .inflate(layout, parent, false)
