+++ WordPress/src/main/java/org/wordpress/android/ui/photopicker/ThumbnailViewUtils.kt
@@ -20,7 +20,7 @@ import org.wordpress.android.util.redirectContextClickToLongPressListener
 import java.util.Locale
 
 @Deprecated("This class is being refactored, if you implement any change, please also update " +
-        "{@link org.wordpress.android.ui.photopicker.mediapicker.ThumbnailViewUtils}")
+        "{@link org.wordpress.android.ui.mediapicker.ThumbnailViewUtils}")
 class ThumbnailViewUtils(val imageManager: ImageManager) {
     fun setupThumbnailImage(
         imgThumbnail: ImageView,
